+++ ubo_app/services/090-web-ui/web-app/src/global.d.ts
@@ -1,0 +1,7 @@
+export {};
+
+declare global {
+  interface Window {
+    GRPC_ENVOY_LISTEN_PORT: string;
+  }
+}